Add user existence check in get_user and delete_user functions

+++ lambda_function.py
@@ -45,6 +45,10 @@ def lambda_handler(event, context):
     return response
 
 def get_user(user_id):
+    # check if user exists
+    existing_user_response = dynamodb_table.get_item(Key={'userid': user_id})
+    if 'Item' not in existing_user_response:
+        return build_response(404, f'User with userid {user_id} not found')
     try:
         response = dynamodb_table.get_item(Key={'userid': user_id})
         return build_response(200, response.get('Item'))
@@ -113,6 +117,10 @@ def modify_user(user_id, update_key, update_value):
         return build_response(400, e.response['Error']['Message'])
 
 def delete_user(user_id):
+    # Check if user exists
+    existing_user_response = dynamodb_table.get_item(Key={'userid': user_id})
+    if 'Item' not in existing_user_response:
+        return build_response(404, f'User with userid {user_id} not found')
     try:
         response = dynamodb_table.delete_item(
             Key={'userid': user_id},
